split clients/mod.rs
