+++ src/utils/mod.rs
@@ -65,3 +65,11 @@ pub fn light_theme_from_colorfgbg(colorfgbg: &str) -> Option<bool> {
     let light = v > 128.0;
     Some(light)
 }
+
+pub fn init_tokio_runtime() -> anyhow::Result<tokio::runtime::Runtime> {
+    use anyhow::Context;
+    tokio::runtime::Builder::new_current_thread()
+        .enable_all()
+        .build()
+        .with_context(|| "Failed to init tokio")
+}
